# PCHR-3084: Single gulpfile for all CiviHR extensions