--- a/hrjobcontract/gulp-tasks/sass.js
+++ b/hrjobcontract/gulp-tasks/sass.js
@@ -6,6 +6,8 @@ module.exports = function () {
      * The app style relies on compass's gems, so we need to rely on it
      * for the time being
      *
+     * @NOTE: will be removed when hrjobcontract is migrated to Shoreditch
+     *
      * @param {Function} done
      */
     main: function (done) {