--- a/hrui/gulp-tasks/requirejs.js
+++ b/hrui/gulp-tasks/requirejs.js
@@ -7,6 +7,10 @@ var uglify = require('gulp-uglify');
 
 module.exports = function () {
   return {
+    /**
+     * Standard JS pipeline without requirejs: uglifies, concats and outputs
+     * a single minified file, while writing a sourcemap in the process
+     */
     main: function () {
       gulp.src(path.join(__dirname, '..', 'js/src/**/*.js'))
         .pipe(sourcemaps.init())
@@ -15,8 +19,13 @@ module.exports = function () {
         .pipe(sourcemaps.write())
         .pipe(gulp.dest(path.join(__dirname, '..', 'js/dist')));
     },
+    /**
+     * Detects if there are any js files in the js/src folder
+     *
+     * @return {Boolean}
+     */
     canRunCriteria: function () {
-      return find.fileSync(/js\/src\/[^/]+.js$/, path.join(__dirname, '..')).length;
+      return find.fileSync(/js\/src\/[^/]+\.js$/, path.join(__dirname, '..')).length;
     }
   };
 };