--- a/org.civicrm.bootstrapcivihr/scss/pages/_vacancies.scss
+++ b/org.civicrm.bootstrapcivihr/scss/pages/_vacancies.scss
@@ -16,7 +16,6 @@
     }
 
     .crm-block.crm-form-block .form-layout-compressed .crm-profile-selector-preview {
-      @extend .btn-primary;
       margin-left: 20px;
 
       .crm-i {