--- a/org.civicrm.reqangular/gulp-tasks/requirejs.js
+++ b/org.civicrm.reqangular/gulp-tasks/requirejs.js
@@ -17,7 +17,7 @@ module.exports = function () {
               moduleSystem: 'RequireJS'
             }))
             .pipe(replace("['angular']", "['common/angular']"))
-            .pipe(replace('module("templates")', 'module("common.templates", [])'))
+            .pipe(replace('module(\'templates\')', 'module(\'common.templates\', [])'))
             .pipe(gulp.dest(path.join(commonFolder, 'modules')));
 
           cb();