--- a/org.civicrm.reqangular/gulp-tasks/requirejs.js
+++ b/org.civicrm.reqangular/gulp-tasks/requirejs.js
@@ -13,9 +13,7 @@ module.exports = function () {
         name: 'cache-templates',
         fn: function (cb) {
           gulp.src(path.join(commonFolder, 'templates', '/**/*.html'))
-            .pipe(templateCache({
-              moduleSystem: 'RequireJS'
-            }))
+            .pipe(templateCache({ moduleSystem: 'RequireJS' }))
             .pipe(replace("['angular']", "['common/angular']"))
             .pipe(replace('module(\'templates\')', 'module(\'common.templates\', [])'))
             .pipe(gulp.dest(path.join(commonFolder, 'modules')));