--- a/uk.co.compucorp.civicrm.hrcore/.gitignore
+++ b/uk.co.compucorp.civicrm.hrcore/.gitignore
@@ -1,4 +1,5 @@
 backstop_data/backstop.temp.json
 backstop_data/site-config.json
+backstop_data/ci_report*
 backstop_data/html_report*
 backstop_data/screenshots