… com.civicrm.hrjobroles/gulpfile.js
… com.civicrm.hrjobroles/package.json
… contactaccessrights/gulpfile.js
… contactaccessrights/package.json
… contactsummary/gulpfile.js
… contactsummary/package.json
… hrjobcontract/gulpfile.js
… hrjobcontract/package.json
… hrui/gulpfile.js
… hrui/package.json
… org.civicrm.bootstrapcivihr/css/civihr.css
… org.civicrm.bootstrapcivihr/gulpfile.js
… org.civicrm.bootstrapcivihr/package.json
… org.civicrm.reqangular/gulpfile.js
… org.civicrm.reqangular/package.json
… org.civicrm.reqangular/src/common/modules/templates.js
… uk.co.compucorp.civicrm.hrcore/gulp/tasks/sass.js
… uk.co.compucorp.civicrm.hrcore/gulp/tasks/test.js
… uk.co.compucorp.civicrm.hrcore/gulp/test.js
… uk.co.compucorp.civicrm.hrcore/gulp/utils.js
… uk.co.compucorp.civicrm.hrcore/gulpfile.js
… uk.co.compucorp.civicrm.hrcore/package.json
… uk.co.compucorp.civicrm.hrleaveandabsences/css/leaveandabsence.css
… uk.co.compucorp.civicrm.hrleaveandabsences/gulpfile.js
… uk.co.compucorp.civicrm.hrleaveandabsences/package.json