## Overview
_(From  https://github.com/civicrm/civihr/pull/2366)_

> Currently each extension has a separate gulpfile, but the tasks are basically always the same (sass, requirejs, test) with some custom logic here and there. This leads to a lot of duplication that is hard to maintain and that will only grow over time

In order to fix this problem, this PR introduces a single gulp setup, in the `uk.co.compucorp.civicrm.hrcore` extension that is able to run the common tasks on any given CiviHR extension, while still accounting for the custom needs that each extension might have

### How it works
There are 4 tasks that can be run on any extension
* `requirejs` 
* `sass`
* `test`
* `build` (runs all the above in sequence)

To run a task, pass the extension name as a `--ext` parameter:
```
gulp requirejs --ext org.civicrm.reqangular
```
It is also possible to pass an alias, instead of the full name, as to save some time. The alias is simply the extension name without the domain bit
```
gulp requirejs --ext reqangular
```

### Custom logic
#### Problem
Certain extension need some custom logic to be ran either before or after the main (=shared) task.

Example: `org.civicrm.bootstrapcivihr` need to perform [transformations](https://github.com/civicrm/civihr/blob/5f8ed2ac191b8f5a45c994871bcbe507c0e29a41/org.civicrm.bootstrapcivihr/gulpfile.js#L22-L29) related to the bootstrap namespace after the dist css file is created

```js
.pipe(postcss([postcssPrefix({
  prefix: bootstrapNamespace + ' ',
  exclude: [
    /^html/, /^body/, /page-civi/, /crm-container/, outsideNamespaceRegExp
  ]
})]))
.pipe(transformSelectors(namespaceRootElements, { splitOnCommas: true }))
.pipe(transformSelectors(removeOutsideNamespaceMarker, { splitOnCommas: true }))
```

Some extensions instead require the entire build logic to be replaced, as the shared logic would not apply to them.

Example `org.civicrm.hrjobcontract` uses [Compass](http://compass-style.org/) to build the css files, without relying on gulp [at all](https://github.com/civicrm/civihr/blob/5f8ed2ac191b8f5a45c994871bcbe507c0e29a41/hrjobcontract/gulpfile.js#L12-L19):
```js
gulp.task('sass', function (done) {
  // The app style relies on compass's gems, so we need to rely on it
  // for the time being
  exec('compass compile', function (_, stdout, stderr) {
    console.log(stdout);
    done();
  });
});
```
while the `org.civicrm.hrui` has a js assets [pipeline](https://github.com/civicrm/civihr/blob/5f8ed2ac191b8f5a45c994871bcbe507c0e29a41/hrui/gulpfile.js#L6-L13) that does not use requirejs:
```js
gulp.task('js-bundle', function () {
  gulp.src('js/src/**/*.js')
    .pipe(sourcemaps.init())
    .pipe(uglify())
    .pipe(concat('hrui.min.js'))
    .pipe(sourcemaps.write())
    .pipe(gulp.dest('js/dist'));
});
```

Some extension need to watch additional files that are not covered by the common tasks.

Example: `org.civicrm.reqangular` needs [to watch](https://github.com/civicrm/civihr/blob/5f8ed2ac191b8f5a45c994871bcbe507c0e29a41/org.civicrm.reqangular/gulpfile.js#L42-L43) these patterns `src/common/templates/**/*.html`, `test/mocks/**/*.js` that are custom

```js
gulp.watch('test/mocks/**/*.js', ['requirejs-bundle-mock']);
gulp.watch('src/common/templates/**/*.html', ['cache-templates', 'requirejs-bundle']);
```

And finally some extensions might have diffent criterias for whether or not a task can be ran on them.

Example: by default the `requirejs` task checks whether a `build.js` file is present, but the hrui extension doesn't have one, although it still requires its js files to be processed

#### Solution
Each extension can have a `gulp-tasks/` folder under which a file with the name of one of the three main tasks can be placed (so `gulp-tasks/sass.js`, `gulp-tasks/requirejs.js`, `gulp-tasks/tasks.js`), and this file will be the source of all the custom logic settings.

The file is expected to export a function that returns an object with one or more of these properties:
* `main` - A function that provides the full replacement of the main task.

   Example: hrjobcontract overrides the default `sass` logic:
    ```js
    return {
      main: function (done) {
        exec('compass compile', { cwd: __dirname + '/../' }, function (_, stdout, stderr) {
          console.log(stdout);
          done();
        });
      }
    };
    ```
* `pre` - An array of tasks to be ran _before_ the main task. Each task is an object with a `name` and a `fn` property. The former is the name that will be assigned to the gulp task, the latter is the actual task logic.

  Example: reqangular caches the templates and build the mock file before the main `requirejs` task is performed
  ```js
  return {
    pre: [
      {
        name: 'cache-templates',
        fn: function (cb) {
          gulp.src(path.join(commonFolder, 'templates', '/**/*.html'))
            .pipe(templateCache({
              moduleSystem: 'RequireJS'
            }))
            .pipe(replace("['angular']", "['common/angular']"))
            .pipe(replace('module(\'templates\')', 'module(\'common.templates\', [])'))
            .pipe(gulp.dest(path.join(commonFolder, 'modules')));

          cb();
        }
      },
      {
        name: 'requirejs:mock',
        fn: function (cb) {
          var buildMocksPath = path.join(__dirname, '..', 'build.mocks.js');

          exec('r.js -o ' + buildMocksPath, function (err, stdout, stderr) {
            err && err.code && console.log(stdout);
            cb();
          });
        }
        }
      }
    ]
  };
  ```
* `post` - An array of tasks to be ran _after_ the main task (see `pre` for details about the task object structure).

  Example: bootstrapcivihr peforms transformations on the min css file after it had been generated

  ```js
  return {
    post: [
      {
        name: 'sass:namespace',
        fn: function () {
          var cssDir = __dirname + '/../' + 'css/';

          return gulp.src(cssDir +  '*.css')
            .pipe(postcss([postcssPrefix({
              prefix: bootstrapNamespace + ' ',
              exclude: [
                /^html/, /^body/, /page-civi/, /crm-container/, outsideNamespaceRegExp
              ]
            })]))
            .pipe(transformSelectors(namespaceRootElements, { splitOnCommas: true }))
            .pipe(transformSelectors(removeOutsideNamespaceMarker, { splitOnCommas: true }))
            .pipe(gulp.dest(cssDir));
        }
      }
    ]
  }
  ```
* `watchPattern` - An array of custom glob patterns that will trigger the task

  Example: requirejs triggers the `requirejs` task in case a template or a mock file has changed

  ```js
  watchPatterns: [
    path.join(commonFolder, '**/*.js'),
    path.join(commonFolder, 'templates/**/*.html'),
    path.join(__dirname, '..', 'test/mocks/**/*.js'),
    '!' + path.join(commonFolder, 'modules/templates.js')
  ]
  ```
* `canRunCriteria` - A function that contains custom logic on which to decide whether the task can be ran on the extension or not

  Example: hrui allows `requirejs` to run not whether if the `build.js` file is present (it doesn't have one), but if there are any *.js files at all under the `src/` folder

  ```js
  canRunCriteria: function () {
    return find.fileSync(/js\/src\/[^/]+.js$/, path.join(__dirname, '..')).length;
  }
  ```

### Running tasks on extensions that do not support them
Not all the tasks can be run on an extension. An extension for example might not have any SASS files, and as such the `sass` task do not apply to them. In case the task is ran on an extension that doesn't support it, the task will be simply skipped
```
$ gulp requirejs --ext bootstrapcivihr
[12:53:00] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[12:53:00] Starting 'requirejs'...
Not eligible for this task, skipping...
[12:53:03] Finished 'requirejs' after 2.34 s
```
This applies also for when the `build` task is run
```
$ gulp build --ext bootstrapcivihr
[12:51:15] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[12:51:15] Starting 'build'...
[12:51:17] Starting 'sass (org.civicrm.bootstrapcivihr)'...
[12:51:18] Starting 'sass:sync (org.civicrm.bootstrapcivihr)'...
[12:51:20] Finished 'sass:sync (org.civicrm.bootstrapcivihr)' after 2.28 s
[12:51:20] Starting 'sass:main (org.civicrm.bootstrapcivihr)'...
[12:51:20] Finished 'sass:main (org.civicrm.bootstrapcivihr)' after 134 ms
[12:51:20] Starting 'sass:namespace'...
[12:51:21] Finished 'sass:namespace' after 206 ms
[12:51:21] Finished 'sass (org.civicrm.bootstrapcivihr)' after 4.14 s
[12:51:21] Starting 'requirejs (org.civicrm.bootstrapcivihr)'...
Not eligible for this task, skipping...
[12:51:21] Finished 'requirejs (org.civicrm.bootstrapcivihr)' after 109 ms
[12:51:21] Starting 'test (org.civicrm.bootstrapcivihr)'...
Not eligible for this task, skipping...
[12:51:21] Finished 'test (org.civicrm.bootstrapcivihr)' after 92 ms
[12:51:21] Finished 'build' after 5.41 s
```

### Improvements
#### Check if requirejs optimizer is installed globally
The `requirejs` task uses the requires [optimizer](http://requirejs.org/docs/optimization.html#download), which needs to be installed globally.

In the past this created confusion as if the module was not found, the task would still look as if it was completed successfully.

A check is now added that throws an error if the module is not found globally
```
$ gulp requirejs --ext contactsummary
[13:00:54] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[13:00:54] Starting 'requirejs'...
[13:00:56] Starting 'requirejs:main (org.civicrm.contactsummary)'...
[13:00:56] 'requirejs:main (org.civicrm.contactsummary)' errored after 869 μs
[13:00:56] Error in plugin "Error"
Message:
    The `requirejs` package is not installed globally (http://requirejs.org/docs/optimization.html#download)
```
#### Support for cross-extension requirejs dependencies
Some extensions requires AMD module from other extensions in order to function, and this dependency is marked with a special syntax in the extension's build.js file.

For example this is the build.js file of `org.civicrm.hrjobcontract` , which has a dependency on some of the angular modules of `uk.co.compucorp.civicrm.hrleaveandabsences`
```js
paths: {
  // ...  
  'leave-absences': '%{uk.co.compucorp.civicrm.hrleaveandabsences}/js/angular/src/leave-absences'
},
```
The new gulp setup is able to detect these dependencies and, upon completing the `requirejs` task on an extension, to automatically compile the JS assets of all the extensions that depend on the one the task was first invoked on.

For example, both the `org.civicrm.contactsummary` and `org.civicrm.hrjobcontract` extensions have a dependency on `uk.co.compucorp.civicrm.hrleaveandabsences`, so when running
```
gulp requirejs --ext hrleaveandabsences`
```
in the output it can be seen that `requirejs ` is run automatically on both `contactsummary` and `hrjobcontract`
```
$ gulp requirejs --ext hrleaveandabsences
[13:11:08] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[13:11:08] Starting 'requirejs'...
[13:11:11] Starting 'requirejs:main (uk.co.compucorp.civicrm.hrleaveandabsences)'...
[13:11:14] Finished 'requirejs:main (uk.co.compucorp.civicrm.hrleaveandabsences)' after 3 s
[13:11:14] Starting 'requirejs:rename'...
[13:11:14] Finished 'requirejs:rename' after 43 ms
[13:11:14] Starting 'requirejs:clean'...
[13:11:14] Finished 'requirejs:clean' after 116 ms
[13:11:14] Starting 'requirejs:dependencies (uk.co.compucorp.civicrm.hrleaveandabsences)'...
/* @NOTE: Running requirejs on contactsummary */
[13:11:15] Starting 'requirejs (org.civicrm.contactsummary)'...
[13:11:16] Starting 'requirejs:main (org.civicrm.contactsummary)'...
[13:11:16] Finished 'requirejs:main (org.civicrm.contactsummary)' after 449 ms
[13:11:16] Starting 'requirejs:dependencies (org.civicrm.contactsummary)'...
[13:11:17] Finished 'requirejs:dependencies (org.civicrm.contactsummary)' after 919 ms
[13:11:17] Finished 'requirejs (org.civicrm.contactsummary)' after 2.35 s
/* @NOTE: Running requirejs on hrjobcontract */
[13:11:17] Starting 'requirejs (org.civicrm.hrjobcontract)'...
[13:11:18] Starting 'requirejs:main (org.civicrm.hrjobcontract)'...
[13:11:19] Finished 'requirejs:main (org.civicrm.hrjobcontract)' after 570 ms
[13:11:19] Starting 'requirejs:dependencies (org.civicrm.hrjobcontract)'...
[13:11:19] Finished 'requirejs:dependencies (org.civicrm.hrjobcontract)' after 910 ms
[13:11:19] Finished 'requirejs (org.civicrm.hrjobcontract)' after 2.48 s
[13:11:19] Finished 'requirejs:dependencies (uk.co.compucorp.civicrm.hrleaveandabsences)' after 5.74 s
[13:11:19] Finished 'requirejs' after 11 s
```

This is performed recursively. So if, in the previous scenario, the `com.civicrm.hrjobroles` extension had a dependency on `org.civicrm.contactsummary`, then it would have been automatically built as well as soon as the task was finished on contactsummary

### Error Handling
#### No extension passed
```
$ gulp requirejs
[13:15:23] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[13:15:23] Starting 'requirejs'...
[13:15:23] 'requirejs' errored after 556 μs
[13:15:23] Error in plugin "Error"
Message:
    Extension name not provided, please provide one, i.e.:
  `gulp some-task --ext org.civicrm.myextension` (with full name)
  `gulp some-task --ext myextension` (with alias)
```
#### No extension passed
```
$ gulp requirejs --ext org.civicrm.foobarbaz
[13:16:46] Using gulpfile ~/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[13:16:46] Starting 'requirejs'...
[13:16:47] 'requirejs' errored after 1.15 s
[13:16:47] Error in plugin "Error"
Message:
    Extension "org.civicrm.foobarbaz" not found
```
#### Error while building (sass, requirejs)
Although a bit brute-force (especially if the task is part of a `build` sequence), it has been necessary to directly kill the node process in case of requirejs or sass errors, after outputting the error
```
$ gulp requirejs --ext contactsummary
[13:19:33] Using gulpfile ~/GitHub/buildkit/build/hr17/sites/all/modules/civicrm/tools/extensions/civihr/uk.co.compucorp.civicrm.hrcore/gulpfile.js
[13:19:33] Starting 'requirejs'...
[13:19:35] Starting 'requirejs:main (org.civicrm.contactsummary)'...

Tracing dependencies for: contact-summary
Error: ENOENT: no such file or directory, open '~/contactsummary/js/src/contact-summary/controllersss/contact-summary.controller.js'
In module tree:
    contact-summary
      contact-summary/modules/contact-summary.module
        contact-summary/modules/contact-summary.controllers

Error: Error: ENOENT: no such file or directory, open '~/contactsummary/js/src/contact-summary/controllersss/contact-summary.controller.js'
In module tree:
    contact-summary
      contact-summary/modules/contact-summary.module
        contact-summary/modules/contact-summary.controllers

    at Object.fs.openSync (fs.js:646:18)
```


## Technical Details
### Update to Node LTS (v8.9.4)
As part of the plan to switch the CiviHR development environment to Node LTS (v8.9.4) from v5.12.0, the npm dependencies in all package.json files (the ones left after this PR at least) had been updgraded, and all extensions had their assets recompiled.

A couple of issues do to the bump in versions had to be fixed:

1. This [`@extend`](https://github.com/civicrm/civihr/pull/2425/commits/89a8d874137bcb308fdde46a0c9acedf004aca7e) was previously ignored, but after the upgrade it was throwing an error, so it was removed
2. The [`common.templates`](https://github.com/civicrm/civihr/blob/staging/org.civicrm.reqangular/src/common/modules/templates.js) module in angular is created on the fly during the build process via the [gulp-angular-templatecache](https://www.npmjs.com/package/gulp-angular-templatecache) module, but it's amended in a couple of places by a simple search and replace before being saved, in order to function correctly with our AMD modules architecture
    ```js
    gulp.src('src/common/templates/**/*.html')
      .pipe(templateCache({
        moduleSystem: 'RequireJS'
      }))
      .pipe(replace("['angular']", "['common/angular']"))
      .pipe(replace('module("templates")', 'module("common.templates", [])'))
      .pipe(gulp.dest('src/common/modules'));
    ```
    After the version bump, the module file that was generated did not use double quotes for strings anymore, but single quotes instead, hence
    ```diff
    - .pipe(replace('module("templates")', 'module("common.templates", [])'))
    + .pipe(replace('module(\'templates\')', 'module(\'common.templates\', [])'))
    ```

### Removed hrjobcontract map files
For some reasons there were several map files in the repo, all in the hrjobcontract extension. They have been deleted now

### Jenkinsfile updates

In order to be able to resolve the path to the extension passed to the `gulp` command via the `--ext`  option, the `cv path` command is [used](https://github.com/civicrm/civihr/blob/0e9c9f8663bcca916f58c70dbc255e36d5e2fa88/uk.co.compucorp.civicrm.hrcore/gulp/utils.js#L230). This command can only find the path of extensions currently installed in the site, so the `Jenkinsfile` had to be updated to make sure the extensions will be installed after the site is built. This was done on https://github.com/civicrm/civihr/pull/2425/commits/0e9c9f8663bcca916f58c70dbc255e36d5e2fa88. We basically call the `drush-install.sh` (the same one used by buildkit) to take care of installing the extensions.

Since now we don't need to install dependencies for each of the extensions, the way we install the JS packages during the build has been updated. First, the structure of the hashmap returned by `listCivihrExtensions()` has been updated, and it now returns something in this format:

```groovy
[
    hrjobroles: [
      name: 'Job Roles',
      folder: 'com.civicrm.hrjobroles',
      hasJSTests: true,
      hasPHPTests: true
    ],
    contactaccessrights: [
      name: 'Contacts Access Rights',
      folder: 'contactaccessrights',
      hasJSTests: true,
      hasPHPTests: true
    ]
]
```

This allows us to have direct access to each of the items in the list, without having to loop through all of them. The main reason for this was to allow us to install the `hrcore` packages before we start the loop to test each extension, as it can be seen [here](https://github.com/civicrm/civihr/blob/2483e651ddee7f2c5da17f86c770686703907642/Jenkinsfile#L124-L128). Next, a `hasJSPackages` property was added to each of the extensions configuration. When it's true, it means that we'll [install the packages for that extension as well, right before running the tests](https://github.com/civicrm/civihr/blob/2483e651ddee7f2c5da17f86c770686703907642/Jenkinsfile#L135-L145). This is necessary because some extensions have custom logic that require a different set of dependencies. Finally, the `testJS()` function [has been updated](https://github.com/civicrm/civihr/blob/2483e651ddee7f2c5da17f86c770686703907642/Jenkinsfile#L323-L337) to use the new `--ext` option. Since now we need to run the tests from within the `hrcore` folder, the function signature was also updated to receive the `hrcore` folder as a param, so that we can cd to it and run the command from the right place.